+++ crates/nostr-relay-pool/src/prelude.rs
@@ -15,6 +15,7 @@ pub use nostr_database::*;
 
 // Internal modules
 pub use crate::policy::*;
+pub use crate::pool::builder::*;
 pub use crate::pool::constants::*;
 pub use crate::pool::options::*;
 pub use crate::pool::{self, *};