+++ crates/nostr-relay-pool/src/relay/mod.rs
@@ -37,7 +37,6 @@ pub use self::options::{
 };
 pub use self::stats::RelayConnectionStats;
 pub use self::status::RelayStatus;
-use crate::policy::AdmitPolicy;
 use crate::shared::SharedState;
 use crate::transport::websocket::{BoxSink, BoxStream};
 
@@ -155,16 +154,6 @@ impl Relay {
         Self {
             inner: AtomicDestructor::new(InnerRelay::new(url, state, opts)),
         }
-    }
-
-    /// Set an admission policy
-    #[inline]
-    pub fn set_admit_policy<T>(&self, policy: T) -> Result<(), Error>
-    where
-        T: AdmitPolicy + 'static,
-    {
-        self.inner.state.set_admit_policy(policy)?;
-        Ok(())
     }
 
     /// Get relay url